+++ src/core/include/mp-units/bits/fmt.h
@@ -143,7 +143,7 @@ template<typename Char>
 template<typename Int>
 [[nodiscard]] constexpr std::make_unsigned_t<Int> to_unsigned(Int value)
 {
-  MP_UNITS_EXPECTS(std::is_unsigned_v<Int> || value >= 0);
+  MP_UNITS_PRECONDITION(std::is_unsigned_v<Int> || value >= 0);
   return static_cast<std::make_unsigned_t<Int>>(value);
 }
 
@@ -217,7 +217,7 @@ MP_UNITS_EXPORT_END
 template<std::forward_iterator It>
 [[nodiscard]] constexpr int parse_nonnegative_int(It& begin, It end, int error_value)
 {
-  MP_UNITS_EXPECTS(begin != end && '0' <= *begin && *begin <= '9');
+  MP_UNITS_PRECONDITION(begin != end && '0' <= *begin && *begin <= '9');
   unsigned value = 0, prev = 0;
   auto pos = begin;
   do {
@@ -277,7 +277,7 @@ template<std::forward_iterator It, typename Handler>
 template<std::forward_iterator It, typename Handler>
 [[nodiscard]] constexpr It parse_arg_id(It begin, It end, Handler& handler)
 {
-  MP_UNITS_EXPECTS(begin != end);
+  MP_UNITS_PRECONDITION(begin != end);
   auto ch = *begin;
   if (ch != '}' && ch != ':') return ::mp_units::detail::do_parse_arg_id(begin, end, handler);
   handler.on_auto();
@@ -324,7 +324,7 @@ template<std::forward_iterator It, typename Char = std::iter_value_t<It>>
 [[nodiscard]] constexpr It parse_dynamic_spec(It begin, It end, int& value, fmt_arg_ref<Char>& ref,
                                               MP_UNITS_STD_FMT::basic_format_parse_context<Char>& ctx)
 {
-  MP_UNITS_EXPECTS(begin != end);
+  MP_UNITS_PRECONDITION(begin != end);
   if ('0' <= *begin && *begin <= '9') {
     const int val = ::mp_units::detail::parse_nonnegative_int(begin, end, -1);
     if (val != -1)
@@ -360,7 +360,7 @@ constexpr int code_point_length(It begin)
 template<std::forward_iterator It, typename Specs>
 [[nodiscard]] constexpr It parse_align(It begin, It end, Specs& specs, fmt_align default_align = fmt_align::none)
 {
-  MP_UNITS_EXPECTS(begin != end);
+  MP_UNITS_PRECONDITION(begin != end);
   auto align = fmt_align::none;
   auto pos = begin + code_point_length(begin);
   if (end - pos <= 0) pos = begin;